Finally get the movement range displayed
