+++ Wartorn/Content/Content.mgcb
@@ -1124,7 +1124,7 @@
 /processorParam:TextureFormat=Color
 /build:sprite/Cursor/Select_Cursor.png
 
-#begin sprite/MapOverlay.png
+#begin sprite/AttackOverlay.png
 /importer:TextureImporter
 /processor:TextureProcessor
 /processorParam:ColorKeyColor=255,0,255,255
@@ -1134,5 +1134,17 @@
 /processorParam:ResizeToPowerOfTwo=False
 /processorParam:MakeSquare=False
 /processorParam:TextureFormat=Color
-/build:sprite/MapOverlay.png
+/build:sprite/AttackOverlay.png
+
+#begin sprite/MoveOverlay.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:sprite/MoveOverlay.png
