+++ Wartorn/GameData/Unit.cs
@@ -28,13 +28,17 @@ namespace Wartorn.GameData
         static Dictionary<UnitPair, int> _DammageTable;
         static Dictionary<UnitType, int> _MovementRange;
         static Dictionary<UnitType, int> _VisionRange;
-        Dictionary<MovementType, Dictionary<TerrainType, int>> _TravelCost;
+        static Dictionary<MovementType, Dictionary<TerrainType, int>> _TravelCost;
 
         public static void Init()
         {
             List<UnitType> unittypes = Enum.GetValues(typeof(UnitType)).Cast<UnitType>().ToList();
             unittypes.Remove(UnitType.None);
-            
+
+            List<MovementType> movementtypes = Enum.GetValues(typeof(MovementType)).Cast<MovementType>().ToList();
+            movementtypes.Remove(MovementType.None);
+
+            List<TerrainType> terraintypes = Enum.GetValues(typeof(TerrainType)).Cast<TerrainType>().ToList();
 
             _DammageTable = new Dictionary<UnitPair, int>();
             foreach (UnitType attacker in unittypes)
@@ -58,150 +62,66 @@ namespace Wartorn.GameData
                 _VisionRange.Add(unittype, 0);
             }
 
-            File.WriteAllText("dmgtable.txt", JsonConvert.SerializeObject(_DammageTable.ToArray(), Formatting.Indented));
-            File.WriteAllText("movementrangetable.txt", JsonConvert.SerializeObject(_MovementRange.ToArray(), Formatting.Indented));
-            File.WriteAllText("visionrangetable.txt", JsonConvert.SerializeObject(_VisionRange.ToArray(), Formatting.Indented));
+            _TravelCost = new Dictionary<MovementType, Dictionary<TerrainType, int>>();
+            foreach (MovementType movetype in movementtypes)
+            {
+                _TravelCost.Add(movetype, new Dictionary<TerrainType, int>());
+                foreach (TerrainType terraintype in terraintypes)
+                {
+                    _TravelCost[movetype].Add(terraintype, int.MaxValue);
+                }
+            }
+
+            Directory.CreateDirectory(@"data/");
+            File.WriteAllText(@"data/dmgtable.txt", JsonConvert.SerializeObject(_DammageTable.ToArray(), Formatting.Indented));
+            File.WriteAllText(@"data/movementrangetable.txt", JsonConvert.SerializeObject(_MovementRange.ToArray(), Formatting.Indented));
+            File.WriteAllText(@"data/visionrangetable.txt", JsonConvert.SerializeObject(_VisionRange.ToArray(), Formatting.Indented));
+
+            File.WriteAllText(@"data/traversecosttable.txt", JsonConvert.SerializeObject(_TravelCost, Formatting.Indented));
         }
 
         public static void Load()
         {
-            string dmgtable = File.ReadAllText("dmgtable.txt");
-            _DammageTable = JsonConvert.DeserializeObject<KeyValuePair<UnitPair, int>[]>(dmgtable).ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
+            _DammageTable = new Dictionary<UnitPair, int>();
+            _MovementRange = new Dictionary<UnitType, int>();
+            _VisionRange = new Dictionary<UnitType, int>();
+            _TravelCost = new Dictionary<MovementType, Dictionary<TerrainType, int>>();
+            
+            string dmgtable = File.ReadAllText(@"data/dmgtable.txt");
+            JsonConvert.DeserializeObject<KeyValuePair<UnitPair, int>[]>(dmgtable).ToList().ForEach(kvp =>
+            {
+                _DammageTable.Add(kvp.Key, kvp.Value);
+            });
 
-            string movrangetable = File.ReadAllText("movementrangetable.txt");
-            _MovementRange = JsonConvert.DeserializeObject<KeyValuePair<UnitType, int>[]>(movrangetable).ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
+            string movrangetable = File.ReadAllText(@"data/movementrangetable.txt");
+            JsonConvert.DeserializeObject<KeyValuePair<UnitType, int>[]>(movrangetable).ToList().ForEach(kvp =>
+            {
+                _MovementRange.Add(kvp.Key, kvp.Value);
+            });
 
-            string visionrangetable = File.ReadAllText("visionrangetable.txt");
-            _VisionRange = JsonConvert.DeserializeObject<KeyValuePair<UnitType, int>[]>(visionrangetable).ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
+            string visionrangetable = File.ReadAllText(@"data/visionrangetable.txt");
+            JsonConvert.DeserializeObject<KeyValuePair<UnitType, int>[]>(visionrangetable).ToList().ForEach(kvp =>
+            {
+                _VisionRange.Add(kvp.Key, kvp.Value);
+            });
+
+            string traversecosttable = File.ReadAllText(@"data/traversecosttable.txt");
+            _TravelCost = JsonConvert.DeserializeObject<Dictionary<MovementType, Dictionary<TerrainType, int>>>(traversecosttable);
         }
 
         /// <summary>
         /// 0 : normal travel,
         /// 1 : troubled travel,
-        /// 2 : cant travel
+        /// horizontal 8 : can not travel
         /// </summary>
         /// <param name="unitType"></param>
         /// <param name="TerrainType"></param>
         /// <returns></returns>
-        public static int CanTravel(UnitType unitType,TerrainType TerrainType)
+        public static int GetTravelCost(UnitType unitType,TerrainType terrainType)
         {
             MovementType movementType = unitType.GetMovementType();
-            int result = 0;
-            switch (TerrainType)
-            {
-                case TerrainType.Reef:
-                    break;
-                case TerrainType.Sea:
-                    break;
-
-                case TerrainType.River:
-                    if (movementType == MovementType.Foot)
-                    {
-                        result = 1;
-                    }
-
-                    break;
-
-                case TerrainType.Coast:
-                    result = 0;
-                    break;
-
-                case TerrainType.Cliff:
-                    result = 2;
-                    break;
-
-                case TerrainType.Road:
-                case TerrainType.Bridge:
-                    result = 0;
-                    break;
-
-                case TerrainType.Plain:
-                    if (movementType == MovementType.Foot || movementType == MovementType.Track)
-                    {
-                        result = 0;
-                    }
-                    else
-                    {
-                        result = 1;
-                    }
-                    break;
-
-                case TerrainType.Tree:
-                    if (movementType == MovementType.Foot)
-                    {
-                        result = 0;
-                    }
-                    else
-                    {
-                        result = 1;
-                    }
-                    break;
-
-                case TerrainType.Mountain:
-                    if (movementType == MovementType.Foot)
-                    {
-                        result = 1;
-                    }
-                    else
-                    {
-                        result = 2;
-                    }
-                    break;
-
-                case TerrainType.MissileSilo:
-                case TerrainType.MissileSiloLaunched:
-                case TerrainType.City:
-                case TerrainType.Factory:
-                case TerrainType.AirPort:
-                case TerrainType.Harbor:
-                case TerrainType.Radar:
-                case TerrainType.SupplyBase:
-                case TerrainType.HQ:
-                    result = 0;
-                    break;
-                default:
-                    break;
-            }
-
-            return result;
+            return _TravelCost[movementType][terrainType];
         }
-
-        public static int CalculateMovementCost(UnitType unitType,TerrainType terrainType)
-        {
-            int result = 1;
-            switch (unitType)
-            {
-                //move by foot
-                case UnitType.Soldier:
-                    break;
-                case UnitType.Mech:
-                    break;
-
-                // move by tires
-                case UnitType.Recon:
-                    break;
-                case UnitType.Rocket:
-                    break;
-                case UnitType.Missile:
-                    break;
-                
-                //move by treads
-                case UnitType.APC:
-                    break;
-                case UnitType.Tank:
-                    break;
-                case UnitType.HeavyTank:
-                    break;
-                case UnitType.Artillery:
-                    break;
-                case UnitType.AntiAir:
-                    break;
-                default:
-                    break;
-            }
-            return result;
-        }
-
         #endregion
 
         #region private field
@@ -259,179 +179,6 @@ namespace Wartorn.GameData
 
     #region JsonConverter class
 
-    public class UnitJsonConverter : JsonConverter
-    {
-        public override bool CanConvert(Type objectType)
-        {
-            return objectType == typeof(Unit);
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            Unit temp = (Unit)value;
-            writer.WriteStartObject();
-            writer.WritePropertyName("UnitType");
-            serializer.Serialize(writer, temp.UnitType.ToString());
-            writer.WritePropertyName("Owner");
-            serializer.Serialize(writer, temp.Owner.ToString());
-            writer.WritePropertyName("HP");
-            serializer.Serialize(writer, temp.HitPoint);
-            writer.WriteEndObject();
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            UnitType unittype = UnitType.None;
-            Owner owner = Owner.None;
-            int hp = 0;
-
-            bool gotUnittype = false;
-            bool gotOwner = false;
-            bool gotHp = false;
-
-            while (reader.Read())
-            {
-                if (reader.TokenType != JsonToken.PropertyName)
-                {
-                    break;
-                }
-
-                string propertyName = (string)reader.Value;
-                if (!reader.Read())
-                {
-                    continue;
-                }
-                switch (propertyName)
-                {
-                    case "UnitType":
-                        unittype = (serializer.Deserialize<string>(reader)).ToEnum<UnitType>();
-                        gotUnittype = true;
-                        break;
-                    case "Owner":
-                        owner = (serializer.Deserialize<string>(reader)).ToEnum<Owner>();
-                        gotOwner = true;
-                        break;
-                    case "HP":
-                        hp = serializer.Deserialize<int>(reader);
-                        gotHp = true;
-                        break;
-                    default:
-                        break;
-                }
-            }
-
-            if (!(gotUnittype && gotHp && gotOwner))
-            {
-                //throw new InvalidDataException("Not enought data");
-                return null;
-            }
-
-            return UnitCreationHelper.Create(unittype, owner, hp);
-        }
-    }
-
-    public class UnitTypeJsonConverter : JsonConverter
-    {
-        public override bool CanConvert(Type objectType)
-        {
-            return objectType == typeof(UnitType);
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            UnitType temp = (UnitType)value;
-            writer.WriteStartObject();
-            writer.WritePropertyName("value");
-            serializer.Serialize(writer, temp.ToString());
-            writer.WriteEndObject();
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            UnitType result = UnitType.None;
-            while (reader.Read())
-            {
-                if (reader.TokenType != JsonToken.PropertyName)
-                {
-                    break;
-                }
-
-                string propertyName = (string)reader.Value;
-                if (!reader.Read())
-                {
-                    continue;
-                }
-
-                if (propertyName == "")
-                {
-                    result = (serializer.Deserialize<string>(reader)).ToEnum<UnitType>();
-                }
-            }
-
-            return result;
-        }
-    }
-
-    public class UnitPairJsonConverter : JsonConverter
-    {
-        public override bool CanConvert(Type objectType)
-        {
-            return objectType == typeof(UnitPair);
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            UnitPair temp = (UnitPair)value;
-
-            writer.WriteStartObject();
-            writer.WritePropertyName("Attacker");
-            serializer.Serialize(writer, temp.Attacker.ToString());
-            writer.WritePropertyName("Defender");
-            serializer.Serialize(writer, temp.Defender.ToString());
-            writer.WriteEndObject();
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            UnitType attacker = default(UnitType);
-            UnitType defender = default(UnitType);
-
-            bool gotAttacker = false;
-            bool gotDefender = false;
-
-            while (reader.Read())
-            {
-                if (reader.TokenType != JsonToken.PropertyName)
-                {
-                    break;
-                }
-
-                string propertyName = (string)reader.Value;
-                if (!reader.Read())
-                {
-                    continue;
-                }
-
-                if (propertyName == "Attacker")
-                {
-                    attacker = (serializer.Deserialize<string>(reader)).ToEnum<UnitType>();
-                    gotAttacker = true;
-                }
-
-                if (propertyName == "Defender")
-                {
-                    defender = (serializer.Deserialize<string>(reader)).ToEnum<UnitType>();
-                    gotDefender = true;
-                }
-            }
-
-            if (!(gotAttacker && gotDefender))
-            {
-                throw new InvalidDataException("Not enought data");
-            }
-
-            return new UnitPair(attacker, defender);
-        }
-    }
+    
     #endregion
 }
